+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Current
 
+* Added MIDI input and output for Windows.
+* Preliminary work on control I/O has begun.
+* Audio testcase is now properly named.
+
 # 0.1.0
 
 * Added `AudioSpecs.toString()`